+++ Athena/Source/Athena/Math/TypesImpl/Vector2.h
@@ -78,7 +78,7 @@ namespace Athena::Math
 
 // -------------Public Methods-------------------------------------
 	public:
-		constexpr SIZE_T Size() const
+		constexpr uint32 Size() const
 		{
 			return Size2;
 		}
@@ -150,13 +150,13 @@ namespace Athena::Math
 
 // -------------Operators-------------------------------------
 	public:
-		constexpr const T& operator[](SIZE_T idx) const
+		constexpr const T& operator[](uint32 idx) const
 		{
 			ATN_CORE_ASSERT(idx < Size2, "Vector subscript out of range");
 			return *(&x + idx);
 		}
 
-		constexpr T& operator[](SIZE_T idx)
+		constexpr T& operator[](uint32 idx)
 		{
 			ATN_CORE_ASSERT(idx < Size2, "Vector subscript out of range");
 			return *(&x + idx);
