+++ Athena/Source/Athena/Math/VectorExponential.h
@@ -7,81 +7,81 @@
 
 namespace Athena::Math
 {
-	template <typename T, SIZE_T Size>
+	template <typename T, uint32 Size>
 	inline Vector<T, Size> Sqrt(const Vector<T, Size>& vec)
 	{
 		Vector<T, Size> out;
-		for (SIZE_T i = 0; i < Size; ++i)
+		for (uint32 i = 0; i < Size; ++i)
 			out[i] = Math::Sqrt(vec[i]);
 
 		return out;
 	}
 
-	template <typename T, SIZE_T Size>
+	template <typename T, uint32 Size>
 	inline Vector<T, Size> Cbrt(const Vector<T, Size>& vec)
 	{
 		Vector<T, Size> out;
-		for (SIZE_T i = 0; i < Size; ++i)
+		for (uint32 i = 0; i < Size; ++i)
 			out[i] = Math::Cbrt(vec[i]);
 
 		return out;
 	}
 
-	template <typename T, SIZE_T Size>
+	template <typename T, uint32 Size>
 	inline Vector<T, Size> ExpE(const Vector<T, Size>& vec)
 	{
 		Vector<T, Size> out;
-		for (SIZE_T i = 0; i < Size; ++i)
+		for (uint32 i = 0; i < Size; ++i)
 			out[i] = Math::ExpE(vec[i]);
 
 		return out;
 	}
 
-	template <typename T, SIZE_T Size>
+	template <typename T, uint32 Size>
 	inline Vector<T, Size> Exp2(const Vector<T, Size>& vec)
 	{
 		Vector<T, Size> out;
-		for (SIZE_T i = 0; i < Size; ++i)
+		for (uint32 i = 0; i < Size; ++i)
 			out[i] = Math::Exp2(vec[i]);
 
 		return out;
 	}
 
-	template <typename T, SIZE_T Size>
+	template <typename T, uint32 Size>
 	inline Vector<T, Size> LogE(const Vector<T, Size>& vec)
 	{
 		Vector<T, Size> out;
-		for (SIZE_T i = 0; i < Size; ++i)
+		for (uint32 i = 0; i < Size; ++i)
 			out[i] = Math::LogE(vec[i]);
 
 		return out;
 	}
 
-	template <typename T, SIZE_T Size>
+	template <typename T, uint32 Size>
 	inline Vector<T, Size> Log2(const Vector<T, Size>& vec)
 	{
 		Vector<T, Size> out;
-		for (SIZE_T i = 0; i < Size; ++i)
+		for (uint32 i = 0; i < Size; ++i)
 			out[i] = Math::Log2(vec[i]);
 
 		return out;
 	}
 
-	template <typename T, SIZE_T Size>
+	template <typename T, uint32 Size>
 	inline Vector<T, Size> Log10(const Vector<T, Size>& vec)
 	{
 		Vector<T, Size> out;
-		for (SIZE_T i = 0; i < Size; ++i)
+		for (uint32 i = 0; i < Size; ++i)
 			out[i] = Math::Log10(vec[i]);
 
 		return out;
 	}
 
-	template <typename T, SIZE_T Size>
+	template <typename T, uint32 Size>
 	inline Vector<T, Size> Pow(const Vector<T, Size>& vec, T exp)
 	{
 		Vector<T, Size> out;
-		for (SIZE_T i = 0; i < Size; ++i)
+		for (uint32 i = 0; i < Size; ++i)
 			out[i] = Math::Pow(vec[i], exp);
 
 		return out;
